task_2/web_app/requirements.txt
Flask
gunicorn==20.1.0
aiohttp==3.6.2
asyncio==3.4.3
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
uvicorn
transformers==4.22.1
huggingface-hub==0.9.1
requests==2.28.1
torch==1.12.1
